+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+## 0.16.0-alpha.0 (2018-3-26)
+
 * Select, DropDownMenu, ColorInput: **improved behaviour on mobile and tablet**.
 * **Breaking**: Remove `hoverable`, `Hoverable`. Recommended alternative: simple state in your component, or CSS-in-JS such as styled-jsx.
 * **Breaking**: Input HOC has changed so that it *always* wraps inputs in a span. This may break some layouts.